# docs/openapi/callpulse-gateways-acl-dialplans.yaml
openapi: 3.0.3
info:
  title: 'FS PBX API v1 — Gateways, Access Controls, Dialplans'
  version: 1.0.0
  description: |
    Contrat des trois ressources ajoutées à l'API v1 de FS PBX pour CallPulse :
    passerelles SIP (`gateways`), listes de contrôle d'accès (`access-controls`)
    et plans de numérotation (`dialplans`).

    Ce document est le contrat de référence : il est écrit avant les contrôleurs
    et fait foi en cas de divergence avec une implémentation.

    ## Authentification

    Toutes les opérations exigent un jeton personnel Sanctum envoyé en
    `Authorization: Bearer <token>`. Deux couches distinctes s'appliquent, dans
    cet ordre, et **elles ne renvoient pas la même forme de corps d'erreur** :

    1. `RequireBearerToken` — en-tête `Authorization` absent → `401`
       `{"success": false, "message": "Bearer token required."}` (forme « middleware »).
    2. `auth:sanctum` — jeton présent mais invalide ou révoqué → `401`
       `{"error": {"type": "authentication_error", ...}}` (forme « enveloppe »).
    3. `AuthorizeUser` — permission manquante → `403` forme « middleware » avec
       `error.code = "forbidden_permission"` ; domaine hors périmètre de
       l'utilisateur → `403` forme « enveloppe » avec `error.code = "forbidden_domain"`.

    Un client doit donc accepter les **deux** formes sur `401` et sur `403`.

    ## Débit

    Limiteur `throttle:api` : **60 requêtes par minute** par utilisateur
...

    ## Codes de retour

    | Code | Signification |
...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  '/api/v1/domains/{domain_uuid}/dialplans':
    get:
      tags: [Dialplans]
      operationId: listDialplans
      summary: 'Lister les plans de numérotation d''un domaine'
      description: |
        Liste paginée **par curseur**, dans l'enveloppe de liste de l'API v1.
        Les éléments ne portent pas leurs lignes de détail : utiliser
        `GET /dialplans/{dialplan_uuid}` pour les obtenir.

        Pagination : `limit` (défaut 50, borné 1–100) et `starting_after`
        (le `dialplan_uuid` du dernier élément de la page précédente).
        **Il n'y a ni `page` ni `per_page`.**

        Permission requise : `dialplan_view`.
      parameters:
        - $ref: '#/components/parameters/DomainUuid'
        - $ref: '#/components/parameters/Limit'
        - name: starting_after
          in: query
          required: false
          description: 'Renvoyer les résultats situés après ce `dialplan_uuid`.'
          schema:
            type: string
            format: uuid
          example: 8c0f2b1d-1f2e-4a3b-9c4d-5e6f7a8b9c0d
      responses:
        200:
          description: Succès.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DialplanListResponse'
        400: { $ref: '#/components/responses/BadRequest' }
        401: { $ref: '#/components/responses/Unauthorized' }
        403: { $ref: '#/components/responses/Forbidden' }
        404: { $ref: '#/components/responses/NotFound' }
        429: { $ref: '#/components/responses/TooManyRequests' }
        500: { $ref: '#/components/responses/InternalError' }

    post:
      tags: [Dialplans]
      operationId: createDialplan
      summary: 'Créer un plan de numérotation et ses lignes de détail'
      description: |
        Écrit le plan (`v_dialplans`) **et** ses lignes de détail
        (`v_dialplan_details`) dans une même transaction, invalide le cache de
        dialplan du contexte, puis émet `bgapi reloadxml` sur l'event socket.

        Le `domain_uuid` de la route est la seule source de vérité ; un
        `domain_uuid` dans le corps est ignoré.

        ## Deux modes d'écriture, exclusifs l'un de l'autre

        - `editor_mode: "builder"` (**défaut**) — le XML FreeSWITCH est **construit
          à partir de `details`**. Un `dialplan_xml` envoyé dans ce mode est ignoré.
        - `editor_mode: "xml"` — le `dialplan_xml` fourni est enregistré tel quel et
          **`details` est ignoré : aucune ligne de détail n'est créée**.

        Choisir le mauvais mode produit un `201` sur un plan vide de détails ou sur
        un XML jamais pris en compte : le mode doit être choisi explicitement dès
        que les deux champs sont envoyés.

        ## Lignes de détail

        Une ligne sans `dialplan_detail_tag` est **silencieusement supprimée** avant
        écriture ; `dialplan_detail_tag` est donc requis pour chaque ligne. Les
        lignes sont triées par `dialplan_detail_group` puis `dialplan_detail_order`
        avant écriture : l'ordre du tableau envoyé n'a pas d'importance, celui de
        ces deux entiers en a.

        ## Garde d'exécution de commande

        Les applications `system`, `bgsystem`, `spawn`, `bg_spawn` et `spawn_stream`
        sont refusées (`400`), en mode builder comme dans le XML brut : elles
        permettraient l'exécution de commandes arbitraires sur le PBX. En mode
        `xml`, le XML est en outre validé syntaxiquement.

        Le XML effectif n'est pas renvoyé — voir le schéma `Dialplan`.

        Permission requise : `dialplan_add`.
      parameters:
        - $ref: '#/components/parameters/DomainUuid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DialplanCreateRequest'
            examples:
              extension_ia:
                summary: 'Extension IA pontée vers une passerelle opérateur'
                value:
                  dialplan_name: callpulse_ai_9001
                  dialplan_number: '9001'
                  dialplan_context: client.exemple.fr
                  dialplan_continue: false
                  dialplan_order: 100
                  dialplan_enabled: true
                  dialplan_description: 'Extension IA CallPulse — poste 9001'
                  editor_mode: builder
                  details:
                    - dialplan_detail_tag: condition
                      dialplan_detail_type: destination_number
                      dialplan_detail_data: '^9001$'
                      dialplan_detail_group: 0
                      dialplan_detail_order: 10
                      dialplan_detail_enabled: true
                    - dialplan_detail_tag: action
                      dialplan_detail_type: export
                      dialplan_detail_data: 'sip_h_X-CallPulse-DID=9001'
                      dialplan_detail_group: 0
                      dialplan_detail_order: 20
                      dialplan_detail_enabled: true
                    - dialplan_detail_tag: action
                      dialplan_detail_type: bridge
                      dialplan_detail_data: 'sofia/gateway/47aa96db-3f70-46ce-bb9a-79ccef396f2b/cp90029ad0b3ae08'
                      dialplan_detail_group: 0
                      dialplan_detail_order: 70
                      dialplan_detail_enabled: true
      responses:
        201:
          description: 'Plan de numérotation créé, avec ses lignes de détail.'
          headers:
            Location:
              description: 'URL du plan créé.'
              schema:
                type: string
                example: '/api/v1/domains/4018f7a3-8e0a-47bb-9f4f-04b1313e0e1b/dialplans/8c0f2b1d-1f2e-4a3b-9c4d-5e6f7a8b9c0d'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dialplan'
        400: { $ref: '#/components/responses/BadRequest' }
        401: { $ref: '#/components/responses/Unauthorized' }
        403: { $ref: '#/components/responses/Forbidden' }
        404: { $ref: '#/components/responses/NotFound' }
        429: { $ref: '#/components/responses/TooManyRequests' }
        500: { $ref: '#/components/responses/InternalError' }
  '/api/v1/domains/{domain_uuid}/dialplans/reload':
    post:
      tags: [Dialplans]
      operationId: reloadDialplans
      summary: 'Recharger le XML des plans de numérotation'
      description: |
        Émet `bgapi reloadxml` sur l'event socket. Aucune écriture en base.

        Une réponse moteur commençant par `-ERR` — event socket injoignable
        compris — produit un `422`. Les écritures de plan (`POST`, `PATCH`,
        `DELETE`) déclenchent déjà un `reloadxml` : cet endpoint sert à rattraper
        un rechargement qui a échoué, signalé par le `switch_response` d'un appel
        précédent.

        Bien que le rechargement soit `bgapi` (asynchrone côté FreeSWITCH), la
        réponse est `200` : le code n'atteste que de la **remise** de la commande.

        Sans corps de requête.

        Note d'implémentation : cette route littérale doit être déclarée **avant**
        `/domains/{domain_uuid}/dialplans/{dialplan_uuid}`.

        Permission requise : `dialplan_view`.
      parameters:
        - $ref: '#/components/parameters/DomainUuid'
      responses:
        200:
          description: 'Commande de rechargement remise au moteur.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DialplanReloadResponse'
        400: { $ref: '#/components/responses/BadRequest' }
        401: { $ref: '#/components/responses/Unauthorized' }
        403: { $ref: '#/components/responses/Forbidden' }
        404: { $ref: '#/components/responses/NotFound' }
        422: { $ref: '#/components/responses/SwitchError' }
        429: { $ref: '#/components/responses/TooManyRequests' }
        500: { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Jeton personnel (Sanctum) d'un utilisateur FS PBX, créé depuis le tableau
        de bord. Un jeton par environnement appelant. L'en-tête
        `Authorization: Bearer <token>` est obligatoire sur toutes les opérations.
  parameters:
    DomainUuid:
      name: domain_uuid
      in: path
      required: true
      description: |
        UUID du domaine (tenant). Doit correspondre à `^[0-9a-fA-F-]{36}$`, sinon
        `400`. Le domaine doit être dans le périmètre de l'utilisateur du jeton,
        sinon `403` (`forbidden_domain`) ; s'il n'existe pas, `404`.
      schema:
        type: string
        format: uuid
      example: 4018f7a3-8e0a-47bb-9f4f-04b1313e0e1b
...
    Limit:
      name: limit
      in: query
      required: false
      description: "Nombre d'éléments à renvoyer. Défaut 50, borné entre 1 et 100."
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
      example: 50
  responses:
...
    NotFound:
      description: |
        Ressource inexistante, ou existante mais rattachée à un autre domaine que
        celui de la route.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              type: invalid_request_error
              message: 'Gateway not found.'
              code: resource_missing
              param: gateway_uuid
              doc_url: 'https://www.fspbx.com/docs/api/v1/errors/'
...
    TooManyRequests:
      description: |
        Débit dépassé : 60 requêtes par minute et par utilisateur authentifié
        (à défaut, par adresse IP). Les en-têtes `Retry-After` et
        `X-RateLimit-Remaining` accompagnent la réponse.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    DialplanListResponse:
      type: object
      description: "Enveloppe de liste de l'API v1, pagination par curseur."
      required: [object, url, has_more, data]
      properties:
        object:
          type: string
          enum: [list]
        url:
          type: string
          example: /api/v1/domains/4018f7a3-8e0a-47bb-9f4f-04b1313e0e1b/dialplans
        has_more:
          type: boolean
          description: "S'il vaut `true`, rappeler l'endpoint avec `starting_after` positionné sur le `dialplan_uuid` du dernier élément reçu."
        data:
          type: array
          items:
            $ref: '#/components/schemas/DialplanSummary'
...